refactor(webfrontend): Chnages according to review
