+++ binder-web-frontend/src/pages/home/components/tables-list/tables-list.component.html
@@ -8,7 +8,7 @@
       <mat-icon class="mat-icon-rtl-mirror">
         {{ treeControl.isExpanded(table) ? 'expand_more' : 'chevron_right' }}
       </mat-icon> 
-      <p class="tables__tree-header-title">{{ 'tablesList.tables' | transloco }}</p>
+      <p>{{ 'tablesList.tables' | transloco }}</p>
     </button>
   </mat-tree-node>
 </mat-tree>
