+++ binder-web-frontend/src/shared/models/tableNode.ts
